Refactor Put token code into separate file. Rust made me put it into a subfolder. Fix infinite loop issue in scan_token() by actually incrementing the current var.
