+++ src/lox_scanner.rs
@@ -1,81 +1,7 @@
-use std::fmt;
 
-pub struct Token {
-    pub token_type: TokenType,
-    pub lexeme: String,
-    pub literal: String,
-    pub line: i32
-}
+mod token;
 
-impl Token {
-    fn new(token_type: TokenType, lexeme: String, literal: String, line: i32) -> Token {
-        Token {
-            token_type,
-            lexeme,
-            literal,
-            line
-        }
-    }
-}
-
-impl fmt::Display for Token {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "Lexeme: {}\nTokenType: {:?}\nLiteral: {}\nLine: {}\n", self.lexeme, self.token_type, self.literal, self.line)
-    }
-}
-
-#[derive(Debug)]
-#[allow(dead_code)] // temp. ignore dead code until these get used.
-#[allow(non_camel_case_types)] // allow non-camel-case typs for tokens.
-pub enum TokenType {
-    // Single-character tokens.
-    LEFT_PAREN,
-    RIGHT_PAREN,
-    LEFT_BRACE,
-    RIGHT_BRACE,
-    COMMA,
-    DOT,
-    MINUS,
-    PLUS,
-    SEMICOLON,
-    SLASH,
-    STAR,
-
-    // One or two character tokens
-    BANG,
-    BANG_EQUAL,
-    EQUAL,
-    EQUAL_EQUAL,
-    GREATER,
-    GREATER_EQUAL,
-    LESS,
-    LESS_EQUAL,
-
-    // Literals
-    IDENTIFIER,
-    STRING,
-    NUMBER,
-
-    // Keywords
-    AND,
-    CLASS,
-    ELSE,
-    FALSE,
-    FUN,
-    FOR,
-    IF,
-    NIL,
-    OR,
-    PRINT,
-    RETURN,
-    SUPER,
-    THIS,
-    TRUE,
-    VAR,
-    WHILE,
-
-    EOF
-}
+use token::{Token, TokenType};
 
 pub struct Scanner {
     pub source: Vec<u8>,
@@ -113,7 +39,7 @@ impl Scanner {
 pub fn scan_tokens(mut scanner: Scanner) -> Result<Vec<Token>, LoxError> {
     while !is_at_end(&scanner) {
         scanner.start = scanner.current;
-        scan_token(&scanner);
+        scan_token(&mut scanner);
     }
 
     scanner.tokens.push(Token::new(TokenType::EOF, String::new(), String::new(), 0));
@@ -125,6 +51,6 @@ fn is_at_end(scanner: &Scanner) -> bool {
     scanner.current >= scanner.source.len()
 }
 
-fn scan_token(_scanner: &Scanner) {
-    // do something
+fn scan_token(scanner: &mut Scanner) {
+    scanner.current += 1;
 }
